rename lbrynet.lbrynet_daemon to lbrynet.daemon
