+++ lbrynet/daemon/Publisher.py
