+++ lbrynet/daemon/auth/util.py
