+++ setup.py
@@ -35,8 +35,8 @@ requires = [
 ]
 
 console_scripts = [
-    'lbrynet-daemon = lbrynet.lbrynet_daemon.DaemonControl:start',
-    'lbrynet-cli = lbrynet.lbrynet_daemon.DaemonCLI:main'
+    'lbrynet-daemon = lbrynet.daemon.DaemonControl:start',
+    'lbrynet-cli = lbrynet.daemon.DaemonCLI:main'
 ]
 
 
@@ -62,7 +62,6 @@ setup(
     long_description=long_description,
     keywords="lbry protocol media",
     license='MIT',
-
     packages=find_packages(base_dir, exclude=['tests']),
     install_requires=requires,
     entry_points={'console_scripts': console_scripts},
