+++ tests/unit/lbrynet_daemon/auth/test_server.py
@@ -2,7 +2,7 @@ import mock
 from twisted.trial import unittest
 
 from tests.mocks import mock_conf_settings
-from lbrynet.lbrynet_daemon.auth import server
+from lbrynet.daemon.auth import server
 
 
 class AuthJSONRPCServerTest(unittest.TestCase):
